+++ data-access/package.json
@@ -10,8 +10,8 @@
     "start:host": "func start",
     "start": "npm-run-all --parallel start:host watch",
     "test": "npx jest --watchAll=true",
-    "gen": "graphql-codegen --config src/graphql/schema/builder/codegen.yml",
-    "gen:watch": "graphql-codegen --config src/graphql/schema/builder/codegen.yml --watch  --silent=false",
+    "gen": "graphql-codegen --config src/routes/http/graphql/schema/builder/codegen.yml",
+    "gen:watch": "graphql-codegen --config src/routes/http/graphql/schema/builder/codegen.yml --watch  --silent=false",
     "postinstall": "serenity-bdd update",
     "prebdd": "npm run bdd:clean && npm run bdd:build",
     "bdd": "cucumber-js --config ./features/support/cucumber.json",
