+++ data-access/src/routes/http/graphql/middleware/permission-middleware.ts
