+++ data-access/src/routes/http/graphql/schema/builder/codegen.yml
@@ -2,11 +2,11 @@
 
 overwrite: true
 schema:
-  - './src/graphql/schema/core/graphql-tools-scalars.ts'
-  - './src/graphql/schema/**/**.graphql'
+  - './src/routes/http/graphql/schema/core/graphql-tools-scalars.ts'
+  - './src/routes/http/graphql/schema/**/**.graphql'
 documents: '../ui-community/src/components/**/**.graphql'
 generates:
-  ./src/graphql/schema/builder/generated.ts:
+  ./src/routes/http/graphql/schema/builder/generated.ts:
     config:
       contextType: '../../init/graphql-context-builder#GraphqlContext'
       useIndexSignature: true
@@ -67,7 +67,7 @@ generates:
     plugins:
       - 'typescript'
       - 'typescript-resolvers'
-  ./src/graphql/schema/builder/graphql.schema.json:
+  ./src/routes/http/graphql/schema/builder/graphql.schema.json:
     plugins:
       - 'introspection'
   ../ui-community/src/generated.tsx:
