+++ data-access/src/routes/http/graphql/schema/interfaces/mongo-subdocument.graphql
