+++ data-access/src/routes/http/graphql/schema/interfaces/mutation-result.graphql
