+++ data-access/src/routes/http/graphql/schema/shared/blob-auth-header.graphql
