+++ data-access/src/routes/http/graphql/schema/types/community.graphql
