+++ data-access/src/routes/http/graphql/schema/types/property.resolvers.ts
