+++ data-access/src/routes/http/graphql/schema/types/service-ticket.graphql
