+++ data-access/src/routes/http/graphql/schema/types/violation-ticket.graphql
